Fix six compile errors; the component now builds for ESP32-S3
`esphome compile examples/epaper_spi_mono.yaml` succeeds against ESPHome
2026.7.4: RAM 30.1%, Flash 8.1%. First time any of this has been built.

Two of these were SILENT failures -- code that reads correctly and does nothing:

1. Backend selection resolved every display to IT8951Backend. The generated C++
   was IT8951Backend::IT8951Backend(EPaperSSD1683*). MockObj.__getattr__
   fabricates a MockObj for any attribute (cpp_generator.py:895) and
   MockObj.__slots__ has no `type`, so display_var.type was a C++ expression
   object, not a MockObjClass, and .inherits_from() on it returned another
   always-truthy expression. This is the same trap as MockObj.__eq__ returning
   an expression rather than a bool, one level deeper -- switching from == to
   inherits_from did not fix it.

   Now resolved from the display's PLATFORM NAME in the validated config, which
   is a plain string. Selecting at codegen time is not the "backend by runtime
   platform name" the plan forbids; nothing is decided on-device.

2. The #ifdef guard added for (3) below sat ahead of every #include, so the
   define from defines.h was not yet visible and each backend .cpp compiled to
   nothing. Surfaced only as `undefined reference to vtable for
   EpaperSPIBackend` at link. defines.h is now included before the guard.

The rest:

3. backend_it8951.cpp could not find it8951.h. ESPHome compiles EVERY file in a
   used component (writer.py copy_src_tree), so the unselected backend built
   anyway and pulled in a driver header absent from the config. Each backend
   .cpp is now gated on a codegen-emitted USE_OPENDISPLAY_* define.
4. OD_COLOR_SCHEME_MONO undeclared: the ColorScheme enum lives in
   opendisplay_structs.h, which protocol_types.h never included.
5. OD_BLE_MANUFACTURER_ID undeclared in protocol.cpp: it was defined in
   ble_transport.h, which the transport-independent MSD builder deliberately
   does not include. Moved to protocol.h.
6. esp_ble_gatt_set_local_mtu is declared in esp_gatt_common_api.h, not
   esp_gatts_api.h. Also fixed %u format specifiers -- uint32_t is
   'long unsigned int' on ESP32.

Also adds refresh: always to the Home Assistant example. external_components
defaults to refresh: 1d, so an HA build keeps rebuilding a clone up to a day old
and replays bugs already fixed upstream; the cache lives in the add-on's own
/data volume and is not reachable from the file editor.

Gitignores .esphome/ (build state created next to example configs) and
examples/secrets.yaml.

+++ components/opendisplay/opendisplay.cpp
@@ -4,6 +4,8 @@
 
 #ifdef USE_ESP32
 #include <esp_gap_ble_api.h>
+// esp_ble_gatt_set_local_mtu() lives here, NOT in esp_gatts_api.h.
+#include <esp_gatt_common_api.h>
 #include <esp_gatts_api.h>
 #include <soc/soc_caps.h>
 
@@ -76,7 +78,8 @@ void OpenDisplayComponent::setup_gatt_() {
   // silently dropped (opendisplay_protocol.h:63-73).
   const esp_err_t merr = esp_ble_gatt_set_local_mtu(OD_BLE_PREFERRED_MTU);
   if (merr != ESP_OK)
-    ESP_LOGW(TAG, "esp_ble_gatt_set_local_mtu(%u) failed: %d", OD_BLE_PREFERRED_MTU, merr);
+    ESP_LOGW(TAG, "esp_ble_gatt_set_local_mtu(%u) failed: %d",
+             static_cast<unsigned>(OD_BLE_PREFERRED_MTU), static_cast<int>(merr));
 
   // 0x2446 is both the service and the characteristic UUID. advertise=true so
   // the service UUID appears in the advertisement alongside our manufacturer
@@ -298,7 +301,8 @@ void OpenDisplayComponent::drain_tx_() {
     this->notify_scratch_.assign(r.data, r.data + r.len);
     this->characteristic_->set_value(std::move(this->notify_scratch_));
     this->characteristic_->notify();
-    ESP_LOGV(TAG, "notify %u bytes (0x%02X 0x%02X)", r.len, r.data[0], r.data[1]);
+    ESP_LOGV(TAG, "notify %u bytes (0x%02X 0x%02X)", static_cast<unsigned>(r.len), r.data[0],
+             r.data[1]);
   }
 #endif
 }
@@ -403,15 +407,16 @@ void OpenDisplayComponent::publish_state_() {
 
 void OpenDisplayComponent::dump_config() {
   ESP_LOGCONFIG(TAG, "OpenDisplay:");
-  ESP_LOGCONFIG(TAG, "  Transfer timeout: %u ms", this->transfer_timeout_ms_);
-  ESP_LOGCONFIG(TAG, "  Refresh timeout: %u ms", this->refresh_timeout_ms_);
+  ESP_LOGCONFIG(TAG, "  Transfer timeout: %u ms", static_cast<unsigned>(this->transfer_timeout_ms_));
+  ESP_LOGCONFIG(TAG, "  Refresh timeout: %u ms", static_cast<unsigned>(this->refresh_timeout_ms_));
   ESP_LOGCONFIG(TAG, "  Config blob: %s (%u bytes)", this->config_valid_ ? "valid" : "INVALID",
                 static_cast<unsigned>(this->config_blob_len_));
   ESP_LOGCONFIG(TAG, "  Security: NONE (open, unauthenticated characteristic)");
   if (this->backend_ != nullptr) {
     const auto &caps = this->backend_->capabilities();
-    ESP_LOGCONFIG(TAG, "  Panel: %ux%u, frame %u bytes, stride %u", caps.width, caps.height,
-                  caps.full_frame_bytes, caps.row_stride_bytes);
+    ESP_LOGCONFIG(TAG, "  Panel: %ux%u, frame %u bytes, stride %u", static_cast<unsigned>(caps.width),
+                  static_cast<unsigned>(caps.height), static_cast<unsigned>(caps.full_frame_bytes),
+                  static_cast<unsigned>(caps.row_stride_bytes));
     ESP_LOGCONFIG(TAG, "  Direct write: %s, PIPE_WRITE: %s", YESNO(caps.supports_direct_write),
                   YESNO(caps.supports_pipe_write));
   }
